requirements.txt
# Django Core
Django==4.2.7
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0

# API Documentation
drf-spectacular==0.27.0

# Database
psycopg2-binary==2.9.9

# Image Processing
Pillow==10.1.0

# CORS & Security
django-cors-headers==4.3.1

# Rate Limiting
django-ratelimit==4.1.0
redis==4.6.0
django-redis==5.4.0

# Celery for async task processing
celery[redis]==5.3.4
django-celery-beat==2.5.0

# Environment
python-dotenv==1.0.0

# Development Tools
django-extensions==3.2.3
Pillow==10.1.0
django-debug-toolbar==4.2.0
ipython==8.17.2

# HTTP Client for n8n integration
requests==2.31.0

# Password Hashing (Argon2)
argon2-cffi==23.1.0

# Twilio for Phone Verification
twilio==8.10.0

# ASGI Support
uvicorn[standard]==0.24.0

